Extract VisibilityConfig from gen_scripts module
